The first chunk `@begin docs 1` through `@end docs 1` parsed.
The parsing stops on the line beginning with the second chunk. I suspect that
there is some issue with parsing new line characters wherein there is a second
new line expected at this location, so parsing stops, but without error, and
hence the matched texted up till now is returned successfully. I suspect this
because the same issue (my grammar defining an incorrect number of new lines to
expect) occurred previously.