Improve LaTeX source, add test make target, and resize document
The document is resized to six inches in width and nine inches in height. This
sizing is inspired by the 'Advanced Introduction to C' book written by Narain
Gehani.

The LaTeX source is improved after reviewing a large part of it, and after
beginning to read 'The Not So Short Introduction to LaTeX'.

A new Make target—test—is included to stop tangling the test file when
unnecessary, and to prepare in a small way for automated testing in the future.