Move the default hook function; document moving the logging function
The default hook function is relocated to the section about generating S-EXP SQL
by processing the LISP data produced form parsing tool syntax with the PEG.

Document the nearby chunk which is partially implemented: that chunk pushes
S-EXP SQL statements to a stack, which should be limited in size by a
customizable variable.

The logging function which was nearby was relocated to a new miscellany and
development-utility section at the end of the document. TODO: this function, and
any other miscellaneous functions, should be moved to a utility file outside the
document; it might be prudent to create a meta-package for development purposes,
ensuring that during development a package of utility functions are available.
If that is done, the program itself must not depend on anything considered a
development utility.